#1135: encapsulate PowerShell profile changes in WindowsHelper
