+++ cli/src/main/java/com/devonfw/tools/ide/os/WindowsHelperImpl.java
@@ -1,11 +1,14 @@
 package com.devonfw.tools.ide.os;
 
+import java.nio.file.Path;
+import java.util.ArrayList;
 import java.util.List;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.devonfw.tools.ide.context.IdeContext;
+import com.devonfw.tools.ide.io.FileAccess;
 import com.devonfw.tools.ide.log.IdeLogLevel;
 import com.devonfw.tools.ide.process.ProcessErrorHandling;
 import com.devonfw.tools.ide.process.ProcessMode;
@@ -20,6 +23,9 @@ public class WindowsHelperImpl implements WindowsHelper {
 
   /** Registry key for the users environment variables. */
   public static final String HKCU_ENVIRONMENT = "HKCU\\Environment";
+
+  public static final String POWERSHELL_CODE_SOURCE_FUNCTIONS =
+      ". \"$env:IDE_ROOT\\_ide\\installation\\functions.ps1\"";
 
   /** Common Windows registry base paths containing (uninstall) information for installed applications (system-wide and per-user). */
   private static final String[] REGISTRY_BASE_PATHS = {
@@ -94,6 +100,18 @@ public class WindowsHelperImpl implements WindowsHelper {
     return new WindowsAppInstallation(version, icon, uninstallString, installLocation);
   }
 
+  @Override
+  public void configurePowerShellProfiles(boolean install) {
+
+    if (!this.context.getSystemInfo().isWindows()) {
+      return;
+    }
+
+    // Windows PowerShell 5.x and PowerShell 7+ have different profile locations.
+    modifyPowerShellProfile("powershell", install);
+    modifyPowerShellProfile("pwsh", install);
+  }
+
   private String findUninstallKey(String appName) {
 
     for (String registryBasePath : REGISTRY_BASE_PATHS) {
@@ -109,7 +127,6 @@ public class WindowsHelperImpl implements WindowsHelper {
       }
     }
     return null;
-
   }
 
   /**
@@ -171,4 +188,96 @@ public class WindowsHelperImpl implements WindowsHelper {
     return i;
   }
 
+  private void modifyPowerShellProfile(String executable, boolean install) {
+
+    Path profilePath = getPowerShellProfilePath(executable);
+    if (profilePath == null) {
+      return;
+    }
+
+    String action = install ? "Configuring" : "Removing";
+    LOG.info("{} IDEasy in {}", action, profilePath);
+
+    FileAccess fileAccess = this.context.getFileAccess();
+    List<String> lines = fileAccess.readFileLines(profilePath);
+
+    if (lines == null && !install) {
+      return;
+    }
+
+    List<String> modifiedLines = modifyPowerShellProfileLines(lines, install);
+
+    Path parent = profilePath.getParent();
+    if (parent != null) {
+      fileAccess.mkdirs(parent);
+    }
+
+    fileAccess.writeFileLines(modifiedLines, profilePath);
+    LOG.debug("Successfully updated PowerShell profile {}", profilePath);
+  }
+
+  private Path getPowerShellProfilePath(String executable) {
+
+    try {
+      ProcessResult result = this.context.newProcess()
+          .executable(executable)
+          .addArgs("-NoProfile", "-Command", "$PROFILE.CurrentUserAllHosts")
+          .run(ProcessMode.DEFAULT_CAPTURE);
+
+      if (!result.isSuccessful()) {
+        LOG.debug("{} is not available or its profile could not be determined.", executable);
+        return null;
+      }
+
+      List<String> output = result.getOut();
+      if (output == null || output.isEmpty()) {
+        LOG.debug("{} returned no PowerShell profile path.", executable);
+        return null;
+      }
+
+      String profilePath = output.stream()
+          .map(String::strip)
+          .filter(line -> !line.isEmpty())
+          .findFirst()
+          .orElse(null);
+
+      if (profilePath == null) {
+        LOG.debug("{} returned no PowerShell profile path.", executable);
+        return null;
+      }
+
+      return Path.of(profilePath);
+
+    } catch (Exception e) {
+      // pwsh is optional. Windows PowerShell normally exists on supported Windows versions.
+      LOG.debug("Could not determine profile for {}: {}", executable, e.getMessage());
+      return null;
+    }
+  }
+
+  List<String> modifyPowerShellProfileLines(List<String> lines, boolean install) {
+
+    List<String> modifiedLines;
+
+    if (lines == null) {
+      modifiedLines = new ArrayList<>();
+    } else {
+      modifiedLines = new ArrayList<>(lines);
+    }
+
+    boolean configured = modifiedLines.stream()
+        .map(String::trim)
+        .anyMatch(POWERSHELL_CODE_SOURCE_FUNCTIONS::equals);
+
+    if (install) {
+      if (!configured) {
+        modifiedLines.add(POWERSHELL_CODE_SOURCE_FUNCTIONS);
+      }
+    } else {
+      modifiedLines.removeIf(
+          line -> line.trim().equals(POWERSHELL_CODE_SOURCE_FUNCTIONS));
+    }
+
+    return modifiedLines;
+  }
 }
