+++ cli/src/test/java/com/devonfw/tools/ide/os/WindowsHelperImplPowerShellTest.java
@@ -1,4 +1,4 @@
-package com.devonfw.tools.ide.tool;
+package com.devonfw.tools.ide.os;
 
 import java.util.List;
 
@@ -8,25 +8,37 @@ import org.junit.jupiter.api.Test;
 import com.devonfw.tools.ide.context.AbstractIdeContextTest;
 import com.devonfw.tools.ide.context.IdeTestContext;
 
-class IdeasyCommandletPowerShellTest extends AbstractIdeContextTest {
+class WindowsHelperImplPowerShellTest extends AbstractIdeContextTest {
 
-  private IdeasyCommandlet commandlet;
+  private WindowsHelperImpl helper;
 
   @BeforeEach
   void setUp() {
     IdeTestContext context = newContext(PROJECT_BASIC);
-    this.commandlet = new IdeasyCommandlet(context);
+    this.helper = new WindowsHelperImpl(context);
+  }
+
+  @Test
+  void testConfigurePowerShellProfilesIsMocked() {
+    IdeTestContext context = newContext(PROJECT_BASIC);
+    WindowsHelperMock helper = (WindowsHelperMock) context.getWindowsHelper();
+
+    helper.configurePowerShellProfiles(true);
+    assertThat(helper.getPowerShellProfilesConfigured()).isTrue();
+
+    helper.configurePowerShellProfiles(false);
+    assertThat(helper.getPowerShellProfilesConfigured()).isFalse();
   }
 
   @Test
   void testInstallIntoMissingPowerShellProfile() {
     // act
     List<String> result =
-        this.commandlet.modifyPowerShellProfileLines(null, true);
+        this.helper.modifyPowerShellProfileLines(null, true);
 
     // assert
     assertThat(result).containsExactly(
-        IdeasyCommandlet.POWERSHELL_CODE_SOURCE_FUNCTIONS);
+        WindowsHelperImpl.POWERSHELL_CODE_SOURCE_FUNCTIONS);
   }
 
   @Test
@@ -36,11 +48,11 @@ class IdeasyCommandletPowerShellTest extends AbstractIdeContextTest {
 
     // act
     List<String> result =
-        this.commandlet.modifyPowerShellProfileLines(lines, true);
+        this.helper.modifyPowerShellProfileLines(lines, true);
 
     // assert
     assertThat(result).containsExactly(
-        IdeasyCommandlet.POWERSHELL_CODE_SOURCE_FUNCTIONS);
+        WindowsHelperImpl.POWERSHELL_CODE_SOURCE_FUNCTIONS);
   }
 
   @Test
@@ -52,13 +64,13 @@ class IdeasyCommandletPowerShellTest extends AbstractIdeContextTest {
 
     // act
     List<String> result =
-        this.commandlet.modifyPowerShellProfileLines(lines, true);
+        this.helper.modifyPowerShellProfileLines(lines, true);
 
     // assert
     assertThat(result).containsExactly(
         "Set-Alias ll Get-ChildItem",
         "$env:TEST = \"value\"",
-        IdeasyCommandlet.POWERSHELL_CODE_SOURCE_FUNCTIONS);
+        WindowsHelperImpl.POWERSHELL_CODE_SOURCE_FUNCTIONS);
   }
 
   @Test
@@ -66,16 +78,16 @@ class IdeasyCommandletPowerShellTest extends AbstractIdeContextTest {
     // arrange
     List<String> lines = List.of(
         "Set-Alias ll Get-ChildItem",
-        IdeasyCommandlet.POWERSHELL_CODE_SOURCE_FUNCTIONS);
+        WindowsHelperImpl.POWERSHELL_CODE_SOURCE_FUNCTIONS);
 
     // act
     List<String> result =
-        this.commandlet.modifyPowerShellProfileLines(lines, true);
+        this.helper.modifyPowerShellProfileLines(lines, true);
 
     // assert
     assertThat(result).containsExactly(
         "Set-Alias ll Get-ChildItem",
-        IdeasyCommandlet.POWERSHELL_CODE_SOURCE_FUNCTIONS);
+        WindowsHelperImpl.POWERSHELL_CODE_SOURCE_FUNCTIONS);
   }
 
   @Test
@@ -83,16 +95,16 @@ class IdeasyCommandletPowerShellTest extends AbstractIdeContextTest {
     // arrange
     List<String> lines = List.of(
         "Set-Alias ll Get-ChildItem",
-        "  " + IdeasyCommandlet.POWERSHELL_CODE_SOURCE_FUNCTIONS + "  ");
+        "  " + WindowsHelperImpl.POWERSHELL_CODE_SOURCE_FUNCTIONS + "  ");
 
     // act
     List<String> result =
-        this.commandlet.modifyPowerShellProfileLines(lines, true);
+        this.helper.modifyPowerShellProfileLines(lines, true);
 
     // assert
     assertThat(result).containsExactly(
         "Set-Alias ll Get-ChildItem",
-        "  " + IdeasyCommandlet.POWERSHELL_CODE_SOURCE_FUNCTIONS + "  ");
+        "  " + WindowsHelperImpl.POWERSHELL_CODE_SOURCE_FUNCTIONS + "  ");
   }
 
   @Test
@@ -100,12 +112,12 @@ class IdeasyCommandletPowerShellTest extends AbstractIdeContextTest {
     // arrange
     List<String> lines = List.of(
         "Set-Alias ll Get-ChildItem",
-        IdeasyCommandlet.POWERSHELL_CODE_SOURCE_FUNCTIONS,
+        WindowsHelperImpl.POWERSHELL_CODE_SOURCE_FUNCTIONS,
         "$env:TEST = \"value\"");
 
     // act
     List<String> result =
-        this.commandlet.modifyPowerShellProfileLines(lines, false);
+        this.helper.modifyPowerShellProfileLines(lines, false);
 
     // assert
     assertThat(result).containsExactly(
@@ -118,12 +130,12 @@ class IdeasyCommandletPowerShellTest extends AbstractIdeContextTest {
     // arrange
     List<String> lines = List.of(
         "Set-Alias ll Get-ChildItem",
-        "  " + IdeasyCommandlet.POWERSHELL_CODE_SOURCE_FUNCTIONS + "  ",
+        "  " + WindowsHelperImpl.POWERSHELL_CODE_SOURCE_FUNCTIONS + "  ",
         "$env:TEST = \"value\"");
 
     // act
     List<String> result =
-        this.commandlet.modifyPowerShellProfileLines(lines, false);
+        this.helper.modifyPowerShellProfileLines(lines, false);
 
     // assert
     assertThat(result).containsExactly(
@@ -140,7 +152,7 @@ class IdeasyCommandletPowerShellTest extends AbstractIdeContextTest {
 
     // act
     List<String> result =
-        this.commandlet.modifyPowerShellProfileLines(lines, false);
+        this.helper.modifyPowerShellProfileLines(lines, false);
 
     // assert
     assertThat(result).containsExactlyElementsOf(lines);
@@ -150,7 +162,7 @@ class IdeasyCommandletPowerShellTest extends AbstractIdeContextTest {
   void testUninstallHandlesMissingPowerShellProfile() {
     // act
     List<String> result =
-        this.commandlet.modifyPowerShellProfileLines(null, false);
+        this.helper.modifyPowerShellProfileLines(null, false);
 
     // assert
     assertThat(result).isEmpty();
@@ -160,13 +172,13 @@ class IdeasyCommandletPowerShellTest extends AbstractIdeContextTest {
   void testUninstallRemovesAllDuplicatePowerShellEntries() {
     // arrange
     List<String> lines = List.of(
-        IdeasyCommandlet.POWERSHELL_CODE_SOURCE_FUNCTIONS,
+        WindowsHelperImpl.POWERSHELL_CODE_SOURCE_FUNCTIONS,
         "Set-Alias ll Get-ChildItem",
-        IdeasyCommandlet.POWERSHELL_CODE_SOURCE_FUNCTIONS);
+        WindowsHelperImpl.POWERSHELL_CODE_SOURCE_FUNCTIONS);
 
     // act
     List<String> result =
-        this.commandlet.modifyPowerShellProfileLines(lines, false);
+        this.helper.modifyPowerShellProfileLines(lines, false);
 
     // assert
     assertThat(result).containsExactly(
